Add some more information to the Cluster config.
I would like to move the EDO towards reconciling the state of the
cluster deployment from a ConfigDB entry, rather than creating the
ConfigDB entry to record what we've done. So:

* Require a cluster name.
* Record the template we used.
* Add an optional array of hosts available on this cluster.

For now these are all filled in statically on cluster creation; I would
like to get the list of hosts and the kubeseal cert updated by something
running on the cluster.